// src/generatedApis/models/marker.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Objects API
 * API for managing objects
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { MarkerInstallPosition } from './marker-install-position';
/**
 * 
 * @export
 * @interface Marker
 */
export interface Marker {
    /**
     * 
     * @type {number}
     * @memberof Marker
     */
    markerId?: number;
    /**
     * 
     * @type {number}
     * @memberof Marker
     */
    subnet: number;
    /**
     * 
     * @type {number}
     * @memberof Marker
     */
    node: number;
    /**
     * 
     * @type {MarkerInstallPosition}
     * @memberof Marker
     */
    installPosition: MarkerInstallPosition;
    /**
     * 
     * @type {boolean}
     * @memberof Marker
     */
    activated?: boolean;
    /**
     * 
     * @type {Date}
     * @memberof Marker
     */
    activationTime?: Date;
}